+++ src/Feature/Sms/Bag/DeleteScheduledSmssBag.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smsapi\Client\Feature\Sms\Bag;
+
+class DeleteScheduledSmssBag
+{
+    /** @var array */
+    public $smsIds;
+
+    public function __construct(array $smsIds)
+    {
+        $this->smsIds = $smsIds;
+    }
+}