+++ src/Feature/Sms/SmsHttpFeature.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace Smsapi\Client\Feature\Sms;
 
 use Smsapi\Client\Feature\Data\DataFactoryProvider;
-use Smsapi\Client\Feature\Sms\Bag\DeleteSmsBag;
+use Smsapi\Client\Feature\Sms\Bag\DeleteScheduledSmssBag;
 use Smsapi\Client\Feature\Sms\Bag\ScheduleSmsBag;
 use Smsapi\Client\Feature\Sms\Bag\ScheduleSmssBag;
 use Smsapi\Client\Feature\Sms\Bag\ScheduleSmsToGroupBag;
@@ -197,10 +197,10 @@ class SmsHttpFeature implements SmsFeature
     }
 
     /**
-     * @param DeleteSmsBag $deleteScheduledSmsBag
+     * @param DeleteScheduledSmssBag $deleteScheduledSmsBag
      * @throws SmsapiClientException
      */
-    public function deleteScheduledSms(DeleteSmsBag $deleteScheduledSmsBag)
+    public function deleteScheduledSms(DeleteScheduledSmssBag $deleteScheduledSmsBag)
     {
         $deleteScheduledSmsBag->schDel = implode(',', $deleteScheduledSmsBag->smsIds);
         unset($deleteScheduledSmsBag->smsIds);