+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `DeleteScheduledSmssBag` that replaces deprecated `DeleteSmsBag`
+
+### Changed
+- `DeleteSmsBag` marked as deprecated
+
 ## [2.3.2] - 2019-06-17
 ### Fixed
 - `SmsFeature::deleteScheduledSms`